requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-cleanup==7.0.0
idna==3.4
Pillow==9.4.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
